moved functions out of lib.py, fixed psycopg2-less install, fixed AMBER20 build script
