+++ src/data/.version
@@ -1,2 +1,2 @@
-benchtool v1.0.2
-Wed Jul 21 10:06:25 CDT 2021
+benchtool v1.0.3
+Wed Jul 28 15:00:50 CDT 2021
