+++ src/data/settings.ini
@@ -4,7 +4,7 @@ build_dir =
 bench_dir = 
 
 [common]
-version = 1.0.2
+version = 1.0.3
 dry_run = True
 debug = True
 timeout = 2
