+++ src/library/template_handler.py
@@ -2,6 +2,7 @@
 import copy
 import glob as gb
 import os
+import re
 import shutil as su
 import sys
 
@@ -86,7 +87,7 @@ class init(object):
     # If the setting in enabled, add the provenance data collection script to the script
     def collect_stats(self, template_obj):
         if self.glob.config['config']['collect_stats']:
-            if self.glob.lib.file_owner(os.path.join(self.glob.stg['utils_path'], "lshw")) == "root":
+            if self.glob.lib.files.file_owner(os.path.join(self.glob.stg['utils_path'], "lshw")) == "root":
                 template_obj.append("\n# Provenance data collection script \n")
                 template_obj.append(os.path.join(self.glob.stg['script_path'], "collect_hw_info") + " " + \
                                     self.glob.stg['utils_path'] + " " + \
@@ -135,6 +136,26 @@ class init(object):
 
         return template_obj
 
+    # Check for unpopulated <<<keys>>> in template file
+    def test_template(self, template_file, template_obj):
+
+        key = "<<<.*>>>"
+        unfilled_keys = [re.search(key, line) for line in template_obj]
+        unfilled_keys = [match.group(0) for match in unfilled_keys if match]
+
+        if len(unfilled_keys) > 0:
+            # Conitue regardless
+            if not self.glob.stg['exit_on_missing']:
+                self.glob.lib.msg.warning("Missing parameters were found in '" + self.glob.lib.rel_path(template_file) + "':" + ", ".join(unfilled_keys))
+                self.glob.lib.msg.warning("'exit_on_missing=False' in settings.ini so continuing anyway...")
+            # Error and exit
+            else:
+               # Write file to disk
+                self.glob.lib.write_list_to_file(template_obj, self.glob.tmp_script)
+                self.glob.lib.msg.error("Missing parameters were found after populating '" + self.glob.lib.rel_path(template_file) + "' and exit_on_missing=True in settings.ini: " + ' '.join(unfilled_keys))
+        else:
+            self.glob.log.debug("All build parameters were filled, continuing")
+
     # Get template files required to constuct build script
     def set_build_files(self):
         # Temp build script
@@ -142,7 +163,7 @@ class init(object):
         self.glob.tmp_script = os.path.join(self.glob.basedir, "tmp." + self.glob.script_file)
 
         if self.glob.stg['build_mode'] == "sched":
-            self.glob.sched['template'] = self.glob.lib.find_exact(self.glob.sched['sched']['type'] + ".template", self.glob.stg['template_path'])
+            self.glob.sched['template'] = self.glob.lib.files.find_exact(self.glob.sched['sched']['type'] + ".template", self.glob.stg['template_path'])
 
         # Get application template file name from cfg, otherwise use cfg_label to look for it
         if self.glob.config['general']['template']:
@@ -151,7 +172,7 @@ class init(object):
             self.glob.config['template'] = self.glob.config['metadata']['cfg_label']
 
         # Search for application template file
-        build_template_search = self.glob.lib.find_partial(self.glob.config['template'], os.path.join(self.glob.stg['template_path'], self.glob.stg['build_tmpl_dir']))
+        build_template_search = self.glob.lib.files.find_partial(self.glob.config['template'], os.path.join(self.glob.stg['template_path'], self.glob.stg['build_tmpl_dir']))
 
         # Error if not found
         if not build_template_search:
@@ -164,7 +185,7 @@ class init(object):
         known_compiler_type = True
         try:
             self.glob.compiler['common'] = self.glob.compiler[self.glob.config['config']['compiler_type']]
-            self.glob.compiler['template'] = self.glob.lib.find_exact(self.glob.stg['compile_tmpl_file'], self.glob.stg['template_path'])
+            self.glob.compiler['template'] = self.glob.lib.files.find_exact(self.glob.stg['compile_tmpl_file'], self.glob.stg['template_path'])
         except:
             known_compiler_type = False
             self.glob.compiler['template'] = None
@@ -213,7 +234,7 @@ class init(object):
 
         # Test for missing parameters
         self.glob.lib.msg.low("Validating template...")
-        self.glob.lib.test_template(self.glob.tmp_script, template_obj)
+        self.test_template(self.glob.tmp_script, template_obj)
 
         # Write populated script to file
         self.glob.lib.msg.low(["Writing template... ", ""])
@@ -227,7 +248,7 @@ class init(object):
     
         # Scheduler template file
         if self.glob.stg['bench_mode'] == "sched":
-            self.glob.sched['template'] = self.glob.lib.find_exact(self.glob.sched['sched']['type'] + ".template", os.path.join(self.glob.stg['template_path'], self.glob.stg['sched_tmpl_dir']))
+            self.glob.sched['template'] = self.glob.lib.files.find_exact(self.glob.sched['sched']['type'] + ".template", os.path.join(self.glob.stg['template_path'], self.glob.stg['sched_tmpl_dir']))
 
         # Set bench template to default, if set in bench.cfg: overload
         if self.glob.config['config']['template']:
@@ -355,7 +376,7 @@ class init(object):
 
         self.glob.lib.msg.low("Validating template...")
         # Test for missing parameters
-        self.glob.lib.test_template(self.glob.tmp_script, template_obj)
+        self.test_template(self.glob.tmp_script, template_obj)
 
         # Write populated script to file
         self.glob.lib.msg.low(["Writing template... ", ""])
